+++ src/main/java/it/gov/pagopa/pu/workflow/wf/ingestionflow/paymentsreporting/wfingestion/PaymentsReportingIngestionWF.java
@@ -6,7 +6,7 @@ import it.gov.pagopa.pu.workflow.wf.ingestionflow.BaseIngestionFlowFileWF;
 
 /**
  * Workflow to ingest PaymentsReporting file
- * @see <a href=https://pagopa.atlassian.net/wiki/spaces/SPAC/pages/1339031693/Classificazione+incassi#3.3.-Rendicontazione>Confluence page</a>
+ * @see <a href=hhttps://pagopa.atlassian.net/wiki/spaces/SPAC/pages/1776648487/Import+Rendicontazioni>Confluence page</a>
  */
 @WorkflowInterface
 public interface PaymentsReportingIngestionWF extends BaseIngestionFlowFileWF {
