+++ src/main/java/it/gov/pagopa/pu/workflow/wf/ingestionflow/receipt/csv/wfingestion/ReceiptIngestionWF.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.wfingestion;
+package it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.csv.wfingestion;
 
 import io.temporal.workflow.WorkflowInterface;
 import io.temporal.workflow.WorkflowMethod;
