+++ src/main/java/it/gov/pagopa/pu/workflow/wf/ingestionflow/receipt/csv/wfingestion/ReceiptIngestionWFImpl.java
@@ -1,10 +1,10 @@
-package it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.wfingestion;
+package it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.csv.wfingestion;
 
 import io.temporal.spring.boot.WorkflowImpl;
 import it.gov.pagopa.payhub.activities.dto.ingestion.receipt.ReceiptIngestionFlowFileResult;
 import it.gov.pagopa.pu.workflow.utilities.TaskQueueConstants;
 import it.gov.pagopa.pu.workflow.wf.ingestionflow.BaseIngestionFlowFileWFImpl;
-import it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.config.ReceiptIngestionWfConfig;
+import it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.csv.config.ReceiptIngestionWfConfig;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationContext;
