+++ src/main/java/it/gov/pagopa/pu/workflow/wf/ingestionflow/receipt/pagopa/wfingestion/ReceiptPagopaIngestionWF.java
@@ -5,7 +5,7 @@ import io.temporal.workflow.WorkflowMethod;
 
 /**
  * Workflow to ingest RT originated from PagoPA
- * @see <a href=https://pagopa.atlassian.net/wiki/spaces/SPAC/pages/1473806561/Ricevute+RT>Confluence page</a>
+ * @see <a href=https://pagopa.atlassian.net/wiki/spaces/SPAC/pages/1473806561/Ricevute+RT+fornite+da+pagoPa>Confluence page</a>
  * */
 @WorkflowInterface
 public interface ReceiptPagopaIngestionWF {
