+++ src/main/java/it/gov/pagopa/pu/workflow/wf/ingestionflow/treasury/opi/wfingestion/TreasuryOpiIngestionWF.java
@@ -5,7 +5,7 @@ import io.temporal.workflow.WorkflowMethod;
 
 /**
  * Workflow to ingest Treasury OPI file
- * @see <a href=https://pagopa.atlassian.net/wiki/spaces/SPAC/pages/1339031693/Classificazione+incassi#3.4.1.-Tesoreria-IUFs>Confluence page</a>
+ * @see <a href=https://pagopa.atlassian.net/wiki/spaces/SPAC/pages/1776484733/Import+Tesoreria>Confluence page</a>
  */
 @WorkflowInterface
 public interface TreasuryOpiIngestionWF {
