+++ src/test/java/it/gov/pagopa/pu/workflow/wf/ingestionflow/receipt/csv/ReceiptIngestionWFClientTest.java
@@ -1,12 +1,12 @@
-package it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt;
+package it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.csv;
 
 import it.gov.pagopa.pu.workflow.dto.generated.WorkflowCreatedDTO;
 import it.gov.pagopa.pu.workflow.service.temporal.WorkflowClientService;
 import it.gov.pagopa.pu.workflow.service.temporal.WorkflowService;
 import it.gov.pagopa.pu.workflow.utilities.TaskQueueConstants;
 import it.gov.pagopa.pu.workflow.utils.TemporalTestUtils;
-import it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.wfingestion.ReceiptIngestionWF;
-import it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.wfingestion.ReceiptIngestionWFImpl;
+import it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.csv.wfingestion.ReceiptIngestionWF;
+import it.gov.pagopa.pu.workflow.wf.ingestionflow.receipt.csv.wfingestion.ReceiptIngestionWFImpl;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
